+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 2.0.0 (IN PROGRESS)
--
+- Added fingerprint rule for EXTENT resource
+
 ## 1.0.1 (03-12-2025)
 - Initial release